requirements.txt
efinance
yfinance
pandas
python-dateutil
